feat: move modules out of 'modules' folder, delete unnecessary DTO and Entity files
