--- src/modules/auth/dto/create-auth.dto.ts
@@ -1,1 +1,0 @@
-export class CreateAuthDto {}
