--- src/modules/auth/dto/update-auth.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateAuthDto } from './create-auth.dto';
-
-export class UpdateAuthDto extends PartialType(CreateAuthDto) {}
