--- src/modules/pets/dto/create-pet.dto.ts
@@ -1,1 +1,0 @@
-export class CreatePetDto {}
