--- src/modules/pets/dto/update-pet.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreatePetDto } from './create-pet.dto';
-
-export class UpdatePetDto extends PartialType(CreatePetDto) {}
