--- src/modules/pets/entities/pet.entity.ts
@@ -1,1 +1,0 @@
-export class Pet {}
