+++ src/users/users.service.ts
@@ -1,0 +1,22 @@
+import { Injectable } from '@nestjs/common';
+import { PrismaService } from 'src/prisma/prisma.service';
+import { Prisma, User } from '@prisma/client';
+
+@Injectable()
+export class UsersService {
+  constructor(private readonly prisma: PrismaService) {}
+
+  async findAll() {
+    const users: User[] = await this.prisma.user.findMany();
+
+    return users;
+  }
+
+  async findOne(id: number) {
+    return this.prisma.user.findUnique({ where: { id } });
+  }
+
+  async create(userData: Prisma.UserCreateInput) {
+    return this.prisma.user.create({ data: userData });
+  }
+}
